+++ src/test/java/dev/thiagogonzaga/thrillhousebot/review/FindingPipelineTest.java
@@ -113,6 +113,7 @@ class FindingPipelineTest {
         new InstructionsResolver.ResolvedInstructions("", ""),
         List.of(),
         "",
+        "",
         List.of(
             new FileDiff("a.java", "modified", 3, 0, 3, ""),
             new FileDiff("b.java", "modified", 2, 0, 2, "")),